Add support for running streamed lambdas in run subcommand
